backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
bcrypt==4.1.1
python-jose==3.3.0
python-multipart==0.0.6
pydantic==2.5.2
email-validator==2.1.0
cloudinary==1.36.0
emergentintegrations
